Allow download centos7 series
The following also uses the same code path that computes the base
channel for downloading a charm, that is also the same code path for
deploying a charm for a given series.
